refactor: 프로젝트 찜 목록 조회 기능을 Bookmark -> Project로 변경하여 순환 참조 문제 해결
