+++ src/main/java/com/wagglex2/waggle/domain/bookmark/repository/BookmarkRepository.java
@@ -16,7 +16,12 @@ public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
     boolean existsByUserIdAndRecruitmentId(Long userId, Long recruitmentId);
 
     /**
-     * 특정 사용자가 북마크한 Recruitment ID를 조회한다.
+     * 특정 사용자가 카테고리별로 찜한 공고 ID 목록을 페이지 단위로 조회한다.
+     *
+     * @param userId   사용자 ID
+     * @param category 공고 카테고리
+     * @param pageable 페이징 정보
+     * @return 사용자가 찜한 공고 ID를 담은 Page 객체
      */
     @Query("""
         SELECT b.recruitment.id FROM Bookmark b
