+++ src/main/java/com/wagglex2/waggle/domain/bookmark/service/BookmarkService.java
@@ -1,6 +1,6 @@
 package com.wagglex2.waggle.domain.bookmark.service;
 
-import com.wagglex2.waggle.domain.project.dto.response.ProjectSummaryResponseDto;
+import com.wagglex2.waggle.domain.common.type.RecruitmentCategory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
@@ -11,6 +11,16 @@ public interface BookmarkService {
     Long createBookmark(Long userId, Long recruitmentId);
 
     /**
+     * 특정 사용자가 카테고리별로 찜한 공고 ID 목록을 페이지 단위로 조회한다.
+     *
+     * @param userId   사용자 ID
+     * @param category 공고 카테고리
+     * @param pageable 페이징 정보
+     * @return 사용자가 찜한 공고 ID를 담은 Page 객체
+     */
+    Page<Long> findBookmarkedRecruitmentIdsByUserId(Long userId, RecruitmentCategory category, Pageable pageable);
+
+    /**
      * 특정 사용자가 특정 공고를 찜했는지 여부를 조회한다.
      *
      * @param userId       사용자 ID
@@ -19,6 +29,5 @@ public interface BookmarkService {
      *         찜하지 않았으면 Optional.empty()를 반환한다.
      */
     Optional<Long> findIdByUserIdAndRecruitmentId(Long userId, Long recruitmentId);
-    Page<ProjectSummaryResponseDto> getBookmarkedProjectsByUserId(Long userId, Pageable pageable);
     void deleteBookmark(Long userId, Long bookmarkId);
 }
